requirements_offline.txt
# Offline model dependencies for DeepSeek and other local models
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.20.0
bitsandbytes>=0.41.0  # For quantization (optional)

# Existing dependencies (if not already installed)
langchain
langchain-openai
langchain-anthropic
langchain-community
langgraph
numpy
scipy
scikit-learn
python-dotenv